% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMR.R
\name{omr}
\alias{omr}
\title{Probabilistic Mendelian Randomization under the Omnigenic Genetic Architecture}
\usage{
omr(ny, nx, num.per, l.j, Z, numCore = 1, two_study = T)
}
\arguments{
\item{ny}{an integer specifying the sample size in outcome study.}

\item{nx}{an integer specifying the sample size in exposure study.}

\item{l.j}{a vector representing the LD Scores calculated from reference study.}

\item{Z}{a matrix containing marginal z-scores, with the first coloum containing z-scores for outcome study and the second coloum containing z-scores for exposure study.}

\item{numCore}{a positive integer specifying the number of cores for parallel computing (default = 1)}

\item{two_study}{whether the study is a two-sample setting, the defult value is TRUE.}

\item{n_ref}{an integer specifying the sample size in reference study.}
}
\value{
A list that contains

\item{alpha}{Estimate of causal effect. Assuming the summary statistics are standardized, \code{alpha} represents increase in mean value of Y in s.d. unit of Y (for continuous outcomes) or log-OR of Y (for binary outcomes) associated with per s.d. unit increase in values of X (for continuous exposures) or values of X changing from 0 to 1 (for binary exposures).}

\item{beta}{Proportion of SNP heritability in the exposure variable}

\item{gamma}{Proportion of SNP horizontal pleiotropy effect in the outcome variable}

\item{z_stat}{Z-statistic for test of the causal effect estimate.}

\item{se_beta}{Standard error of causal effect estimate.}

\item{pvalue}{P value for the causal effect, based on the z test.}
}
\description{
Two-sampling Mendelian randomization analysis under omnigenic genetic architecture assumption.
}
\details{
OMR makes an omnigenic modeling assumption on the SNP effects on the exposure variable and uses genome-wide SNPs as instrumental variables without pre-selection to enable more powerful MR method.
}
\examples{
data(exampledata)
attach(exampledata)
res=omr(ny,nx,num.per,l.j,Z)
closeAllConnections()
detach(exampledata)
}
\author{
Lu Wang, Boran Gao, Yue Fun, and Xiang Zhou
}
